Refactored but not manually tested yet
